ECS - Add Terraform test to verify mongo creation and status
There is an EBS volume resource that is not part of Terraform state because it is provisioned by rexray/ebs volume plugin.

Used Go's aws sdk to remove this volume after Terraform destroy. In case of multiple test volumes with same name, the test case fails with an error and the volume must be deleted manually.